Move interfaces to package
