+++ project/app/src/main/java/com/chalmers/tda367/localfeud/data/handler/DataHandlerFacade.java
@@ -1,5 +1,12 @@
 package com.chalmers.tda367.localfeud.data.handler;
 
+import com.chalmers.tda367.localfeud.data.handler.interfaces.IChatDataHandler;
+import com.chalmers.tda367.localfeud.data.handler.interfaces.IChatMessageDataHandler;
+import com.chalmers.tda367.localfeud.data.handler.interfaces.ICommentDataHandler;
+import com.chalmers.tda367.localfeud.data.handler.interfaces.ILikeDataHandler;
+import com.chalmers.tda367.localfeud.data.handler.interfaces.IMeDataHandler;
+import com.chalmers.tda367.localfeud.data.handler.interfaces.IPostDataHandler;
+
 /**
  * Created by ericwenn on 5/12/16.
  */
