--- project/app/src/main/java/com/chalmers/tda367/localfeud/data/handler/DataResponseListener.java
@@ -1,9 +1,0 @@
-package com.chalmers.tda367.localfeud.data.handler;
-
-/**
- * Created by ericwenn on 5/12/16.
- */
-interface DataResponseListener<D> {
-    void onSuccess( D data );
-    void onFailure( DataResponseError error, String errormessage );
-}
