+++ project/app/src/main/java/com/chalmers/tda367/localfeud/data/handler/interfaces/ILikeDataHandler.java
@@ -1,4 +1,4 @@
-package com.chalmers.tda367.localfeud.data.handler;
+package com.chalmers.tda367.localfeud.data.handler.interfaces;
 
 import com.chalmers.tda367.localfeud.data.Like;
 import com.chalmers.tda367.localfeud.data.Post;
